Fix empty pull request and implement TMDb movie search functionality
This commit implements a movie search feature using The Movie Database API. It includes the following changes: Define Codable structures (MovieData and Movie) to represent the JSON response from TMDb API Create a searchMovie function that takes a movie name as input, constructs the API request URL, and performs the API request using URLSession Implement a parseJSON function that decodes the JSON data using JSONDecoder into MovieData and extracts an array of Movie instances Update the searchMovie function to use the parseJSON function for parsing the JSON response and print the titles of the movies in the search results With these changes, users can search for movies by name and retrieve a list of matching titles from TMDb. Note: There's no proper user interface at the moment, only a button to test the functionalities above and print out the results on the console.